# internal/transport/http/apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "企业 AI Coding 能力治理平台。管理接口先登录获取 data.token，再在 Authorize 中填写 AdminBearer（Bearer + 空格 + Token）。Gateway 使用成员 Access Key，与管理员 Token 分开。JSON 中的业务 ID 使用字符串。",
        "title": "Zentrola API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/models": {
            "get": {
                "security": [
                    {
                        "AdminBearer": []
                    }
                ],
                "description": "按 ID 倒序分页，最新模型在前；可通过 status 只查询启用或停用模型；将 nextCursor 作为下一次请求的 after。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模型与资源"
                ],
                "summary": "逻辑模型列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "上一页 nextCursor，默认从头查询",
                        "name": "after",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 50,
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "ACTIVE",
                            "DISABLED"
                        ],
                        "type": "string",
                        "description": "模型状态",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.PageResponse-management_Model"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "请求追踪 ID"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AdminBearer": []
                    }
                ],
                "description": "官方名称、编码及非空输入输出类型必填；新模型默认停用，接入映射单独配置。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模型与资源"
                ],
                "summary": "创建官方模型",
                "parameters": [
                    {
                        "description": "模型字段",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/management.ModelInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/management.Model"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    }
                }
            }
        },
...
        "/v1/chat/completions": {
            "post": {
                "security": [
                    {
                        "GatewayBearer": []
                    }
                ],
                "description": "使用成员 Access Key 的 Bearer 认证。请求体示例：{\"model\":\"deepseek-chat\",\"messages\":[{\"role\":\"user\",\"content\":\"Hello\"}],\"stream\":false}。model 从 GET /v1/models 获取，原生扩展字段透传；stream=true 返回 SSE，建议用 curl 或 SDK 验证流式响应。执行会调用真实上游并可能产生费用。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "text/event-stream"
                ],
                "tags": [
                    "OpenAI Gateway"
                ],
                "summary": "OpenAI Chat Completions",
                "parameters": [
                    {
                        "description": "OpenAI 原生 Chat Completions 请求；至少包含 model、messages",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "上游原生响应；stream=true 时为 SSE 事件流",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "平台请求追踪 ID"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.OpenAIErrorResponse"
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "平台请求追踪 ID"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.OpenAIErrorResponse"
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "平台请求追踪 ID"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/http.OpenAIErrorResponse"
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "平台请求追踪 ID"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.OpenAIErrorResponse"
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "平台请求追踪 ID"
                            }
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large",
                        "schema": {
                            "$ref": "#/definitions/http.OpenAIErrorResponse"
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "平台请求追踪 ID"
                            }
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "schema": {
                            "$ref": "#/definitions/http.OpenAIErrorResponse"
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "平台请求追踪 ID"
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/http.OpenAIErrorResponse"
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "平台请求追踪 ID"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/http.OpenAIErrorResponse"
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "平台请求追踪 ID"
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/http.OpenAIErrorResponse"
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "平台请求追踪 ID"
                            }
                        }
                    },
                    "504": {
                        "description": "Gateway Timeout",
                        "schema": {
                            "$ref": "#/definitions/http.OpenAIErrorResponse"
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "平台请求追踪 ID"
                            }
                        }
                    }
                }
            }
        },
        "/v1/models": {
            "get": {
                "security": [
                    {
                        "GatewayBearer": []
                    }
                ],
                "description": "返回当前成员被授权且有可用 OpenAI 资源的逻辑模型；不接受查询参数。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OpenAI Gateway"
                ],
                "summary": "可用的 OpenAI 逻辑模型",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.OpenAIModelsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.OpenAIErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.OpenAIErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/http.OpenAIErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "gateway.Model": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "owned_by": {
                    "type": "string"
                }
            }
        },
        "github_com_zentrola_zentrola_internal_transport_http.response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                }
            }
        },
...
        "http.OpenAIErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string",
                            "example": "UNAUTHENTICATED"
                        },
                        "message": {
                            "type": "string",
                            "example": "Authentication failed."
                        },
                        "param": {},
                        "type": {
                            "type": "string",
                            "example": "authentication_error"
                        }
                    }
                }
            }
        },
        "http.OpenAIModelsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gateway.Model"
                    }
                },
                "object": {
                    "type": "string",
                    "example": "list"
                }
            }
        },
...
        "http.PageResponse-management_Model": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/management.Model"
                    }
                },
                "nextCursor": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "management.Model": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "0"
                },
                "inputModalities": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "outputModalities": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "management.ModelInput": {
            "type": "object",
            "required": [
                "code",
                "inputModalities",
                "name",
                "outputModalities"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "inputModalities": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "TEXT",
                            "IMAGE",
                            "AUDIO",
                            "VIDEO"
                        ]
                    }
                },
                "name": {
                    "type": "string"
                },
                "outputModalities": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "TEXT",
                            "IMAGE",
                            "AUDIO",
                            "VIDEO"
                        ]
                    }
                },
                "remark": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminBearer": {
            "description": "管理员登录返回的 Token，填写：Bearer \u003ctoken\u003e。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "GatewayBearer": {
            "description": "成员 Access Key，填写：Bearer \u003caccess-key\u003e。可用于 OpenAI 或 Anthropic 接口。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "GatewayKey": {
            "description": "成员 Access Key 原文，仅用于 Anthropic 接口；不要同时发送 Authorization。",
            "type": "apiKey",
            "name": "x-api-key",
            "in": "header"
        }
    }
}